fix issue while logging in
