+++ lib/screens/auth_screen.dart
@@ -1,19 +1,46 @@
 import 'package:flutter/material.dart';
 
-import './signup_screen.dart';
 import '../widgets/custom_text_field.dart';
 import '../services/auth.dart';
 
-class LoginScreen extends StatefulWidget {
-  static const routeName = '/login';
+class AuthScreen extends StatefulWidget {
+  static const routeName = '/auth';
 
-  const LoginScreen({super.key});
+  const AuthScreen({super.key});
 
   @override
-  State<LoginScreen> createState() => _LoginScreenState();
+  State<AuthScreen> createState() => _AuthScreenState();
 }
 
-class _LoginScreenState extends State<LoginScreen> {
+class _AuthScreenState extends State<AuthScreen> {
+  bool _isLogin = true;
+
+  void switchAuthMethod() {
+    setState(() {
+      _isLogin = !_isLogin;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    if (_isLogin) {
+      return _LoginScreen(switchAuthMethod);
+    } else {
+      return _SignUpScreen(switchAuthMethod);
+    }
+  }
+}
+
+class _LoginScreen extends StatefulWidget {
+  final VoidCallback switchMethod;
+
+  const _LoginScreen(this.switchMethod);
+
+  @override
+  State<_LoginScreen> createState() => _LoginScreenState();
+}
+
+class _LoginScreenState extends State<_LoginScreen> {
   final _emailController = TextEditingController();
   final _passwordController = TextEditingController();
   bool isLoading = false;
@@ -90,13 +117,76 @@ class _LoginScreenState extends State<LoginScreen> {
               child: Text('Sign In with Google'),
             ),
             TextButton(
-              onPressed: () {
-                Navigator.pushReplacementNamed(
-                  context,
-                  SignUpScreen.routeName,
-                );
-              },
+              onPressed: widget.switchMethod,
               child: Text('Create an Account'),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class _SignUpScreen extends StatefulWidget {
+  final VoidCallback switchMethod;
+
+  const _SignUpScreen(this.switchMethod);
+
+  @override
+  State<_SignUpScreen> createState() => _SignUpScreenState();
+}
+
+class _SignUpScreenState extends State<_SignUpScreen> {
+  final _emailController = TextEditingController();
+  final _passwordController = TextEditingController();
+  bool isLoading = false;
+
+  void _signUp() async {
+    setState(() => isLoading = true);
+    final auth = AuthServices();
+
+    final scaffoldMessenger = ScaffoldMessenger.of(context);
+
+    final errorMessage = await auth.signUp(
+      _emailController.text,
+      _passwordController.text,
+    );
+
+    if (errorMessage != null) {
+      scaffoldMessenger.showSnackBar(
+        SnackBar(content: Text(errorMessage)),
+      );
+    }
+    setState(() => isLoading = false);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text('Sign Up')),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          children: [
+            TextField(
+              controller: _emailController,
+              decoration: InputDecoration(labelText: 'Email'),
+            ),
+            TextField(
+              controller: _passwordController,
+              obscureText: true,
+              decoration: InputDecoration(labelText: 'Password'),
+            ),
+            const SizedBox(height: 16),
+            isLoading
+                ? CircularProgressIndicator()
+                : ElevatedButton(
+                    onPressed: _signUp,
+                    child: Text('Sign Up'),
+                  ),
+            TextButton(
+              onPressed: widget.switchMethod,
+              child: Text('Login'),
             ),
           ],
         ),
